+++ .storybook/main.js
@@ -1,7 +1,7 @@
 const path = require('path');
 
 module.exports = {
-  stories: ['../src/**/*.stories.(js|mdx)'],
+  stories: ['../src/**/*.stories.(js|mdx)', './welcome-page/welcome.stories.js'],
   addons: [
     '@storybook/addon-actions/register',
     '@storybook/addon-knobs/register',
@@ -18,7 +18,6 @@ module.exports = {
         test: /\.scss$/, use: ['style-loader', 'css-loader', 'sass-loader']
       }
     );
-  
     config.resolve = {
       alias: {
         helpers: path.resolve(__dirname, '__helpers__/')