+++ .storybook/welcome-page/loves-carbon/index.js
@@ -1,0 +1,1 @@
+export { default } from './loves-carbon.component';